Fixed minimum peer requirements to forge.
Default value of 20; when the default value was changed, it would not take affect regardless. Network now authorizes forging regardless of connected peers. Blocks are verified upon next connection to peer, if the node reconnects and has forged its own fork of the chain, it's blocks will be deemed invalid, and it will re-download the chain from connected peers.